+++ internal/acctest/config/passwordcredentialvalidator/ping_one_for_enterprise_directory_password_credential_validator_test.go
@@ -106,7 +106,6 @@ func testAccPingOneForEnterpriseDirectoryPasswordCredentialValidators(resourceNa
 					name = "example"
 				}
 			]
-			inherited = false
 	  	}
 		`
 	}